requirements.txt
asgiref==3.6.0
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.1.0
ci-info==0.3.0
click==8.1.3
configobj==5.0.8
configparser==5.3.0
cryptography==40.0.2
Django==4.2
django-cors-headers==3.14.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
requests==2.28.2
docopt==0.6.2
etelemetry==0.3.0
filelock==3.11.0
future==0.18.3
httplib2==0.22.0
idna==3.4
isodate==0.6.1
looseversion==1.1.2
lxml==4.9.2
networkx==3.1
nibabel==5.1.0
nipype==1.8.6
numpy==1.24.2
packaging==23.1
pandas==2.0.0
pathlib==1.0.1
PDFNetPython3==9.5.0.post1
pdfrw==0.4
prov==2.0.0
psycopg2-binary==2.9.6
pyasn1==0.4.8
pycparser==2.21
pycryptodome==3.17
pydot==1.4.2
PyJWT==2.6.0
PyMuPDF==1.22.0
pyOpenSSL==23.1.1
pyparsing==3.0.9
PyPDF2==1.25.1
PyPDF4==1.27.0
python-dateutil==2.8.2
pytz==2023.3
pyxnat==1.5
rdflib==6.3.2
requests==2.28.2
rsa==4.9
scipy==1.10.1
simplejson==3.19.1
six==1.16.0
sqlparse==0.4.3
traits==6.3.2
tzdata==2023.3
urllib3==1.26.15
